Temporarily disabled the android to watch push of data and introduced nanohttpd
Details about the issue can be found here:
https://forums.garmin.com/developer/connect-iq/f/legacy-bug-reports/5144/failure_during_transfer-issue-again-now-using-comm-sample
People have reported receiving a FAILURE_DURING_TRANSFER when sending messages to the watch with a short interval, this also
seems to happen after a certain number of messages have been sent to the watch according to people on the forum. On my end
two people one with a F5X+ and another with a VivoActive 3 Music have reporting the issue intermittently too. So instead of
having the android app push the data I am now having the watch poll the data via a /data request to localhost on an ephemeral
port. This is a workaround until Garmin decides to tackle the underlying issue (the forum post is from 2yrs+ ago)